+++ Modules/commands/_hooks.py
@@ -79,40 +79,3 @@ def hook(name: str):
     return real_decorator
 
 
-@hook("require_channel")
-class RequireChannelHook(HookImplementation):
-    """
-    Require Channel hook implementation
-    """
-    __slots__ = ['message']
-
-    def __init__(self, message="this MUST be executed in a channel"):
-        self.message = message
-
-    def __repr__(self):
-        return f"RequireChannelHook(message='{self.message}')"
-
-    async def pre_execute(self, context: Context):
-        LOG.debug("in require_channel")
-        if not context.channel:
-            await context.reply(self.message)
-            raise self.Cancel(self.message)
-
-
-@hook("require_dm")
-class RequireDirectMessageHook(HookImplementation):
-    __slots__ = ['message']
-
-    def __init__(self, message="this MUST be executed from a direct message with me!"):
-        self.message = message
-
-    def __repr__(self):
-        return f"RequireDirectMessageHook(message='{self.message}')"
-
-    async def pre_execute(self, context: Context):
-        LOG.debug("in RequireDirectMessage setup...")
-        if context.channel:
-            await context.reply(self.message)
-            raise self.Cancel(self.message)
-
-
